ai/langchain/demo_langgraph/requirements.txt
black>=25.1.0
isort>=6.0.1
langchain>=0.3.25
langchain-openai>=0.3.16
langchain-community>=0.3.23
langgraph>=0.4.3
openai>=1.78.0
mypy>=1.15.0
python-dotenv>=1.1.0
tavily-python>=0.7.2